# public/.well-known/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "IQ Auto Deals Inventory API",
    "description": "Search vehicles on IQ Auto Deals, a nationwide online car marketplace connecting buyers with certified dealers across all 50 US states. Buyers can compare vehicles and receive competing offers from dealers.",
    "version": "1.0.0",
    "contact": {
      "name": "IQ Auto Deals",
      "url": "https://iqautodeals.com",
      "email": "support@iqautodeals.com"
    }
  },
  "servers": [{ "url": "https://iqautodeals.com", "description": "Production" }],
  "paths": {
    "/api/ai-inventory": {
      "get": {
        "operationId": "searchVehicles",
        "summary": "Search vehicle inventory",
        "description": "Search for vehicles by make, model, year, price, location, body type, fuel type, drivetrain, mileage, and condition.",
        "parameters": [
          { "name": "q", "in": "query", "schema": { "type": "string" }, "description": "Free-text search (e.g., 'reliable 3rd row', 'ram 5500', 'awd')", "example": "reliable suv 3rd row" },

          { "name": "make", "in": "query", "schema": { "type": "string" }, "description": "Vehicle make (e.g., Toyota, Ford, BMW, Mercedes-Benz)", "example": "Toyota" },
          { "name": "model", "in": "query", "schema": { "type": "string" }, "description": "Vehicle model (e.g., Camry, F-150, 3 Series)", "example": "Camry" },

          { "name": "yearMin", "in": "query", "schema": { "type": "integer" }, "description": "Minimum model year", "example": 2020 },
          { "name": "yearMax", "in": "query", "schema": { "type": "integer" }, "description": "Maximum model year", "example": 2026 },

          { "name": "minPrice", "in": "query", "schema": { "type": "number" }, "description": "Minimum price in USD", "example": 10000 },
          { "name": "maxPrice", "in": "query", "schema": { "type": "number" }, "description": "Maximum price in USD", "example": 35000 },

          { "name": "lat", "in": "query", "schema": { "type": "number" }, "description": "Latitude for proximity search", "example": 33.749 },
          { "name": "lng", "in": "query", "schema": { "type": "number" }, "description": "Longitude for proximity search", "example": -84.388 },
          { "name": "radiusMiles", "in": "query", "schema": { "type": "number", "default": 100 }, "description": "Search radius in miles (default 100)", "example": 50 },

          { "name": "state", "in": "query", "schema": { "type": "string" }, "description": "US state abbreviation (e.g., GA, FL, TX, CA)", "example": "GA" },
          { "name": "city", "in": "query", "schema": { "type": "string" }, "description": "City name (partial match supported)", "example": "Atlanta" },

          {
            "name": "bodyType",
            "in": "query",
            "schema": { "type": "string", "enum": ["SUV", "Sedan", "Truck", "Coupe", "Hatchback", "Convertible", "Minivan", "Wagon"] },
            "example": "SUV"
          },
          {
            "name": "fuelType",
            "in": "query",
            "schema": { "type": "string", "enum": ["Gasoline", "Diesel", "Hybrid", "Electric", "Plug-In Hybrid"] },
            "example": "Gasoline"
          },
          {
            "name": "drivetrain",
            "in": "query",
            "schema": { "type": "string", "enum": ["AWD", "FWD", "RWD", "4WD"] },
            "example": "AWD"
          },
          {
            "name": "condition",
            "in": "query",
            "schema": { "type": "string", "enum": ["New", "Used", "Certified Pre-Owned"] },
            "example": "Used"
          },

          { "name": "maxMileage", "in": "query", "schema": { "type": "integer" }, "description": "Maximum mileage", "example": 50000 },

          {
            "name": "sort",
            "in": "query",
            "schema": { "type": "string", "enum": ["relevance", "distance", "priceAsc", "priceDesc", "yearDesc", "mileageAsc"], "default": "relevance" },
            "description": "Sort order (distance requires lat/lng)"
          },

          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 20, "maximum": 50 }, "description": "Number of results to return (max 50, default 20)" },
          { "name": "offset", "in": "query", "schema": { "type": "integer", "default": 0 }, "description": "Number of results to skip for pagination" }
        ],
        "responses": {
          "200": {
            "description": "Vehicle search results",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["vehicles", "total", "limit", "offset"],
                  "properties": {
                    "vehicles": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": ["title", "image_url", "vin", "year", "make", "model", "price", "listing_url"],
                        "properties": {
                          "title": { "type": "string", "description": "Full vehicle title (Year Make Model Trim)" },
                          "image_url": { "type": ["string", "null"], "format": "uri", "description": "Direct URL to the main vehicle photo (JPEG). Display this image for each vehicle result." },
                          "primaryImage": { "type": ["string", "null"], "format": "uri", "description": "Same as image_url. Direct URL to vehicle photo." },
                          "vin": { "type": "string" },
                          "year": { "type": "integer" },
                          "make": { "type": "string" },
                          "model": { "type": "string" },
                          "trim": { "type": ["string", "null"] },
                          "bodyType": { "type": ["string", "null"] },
                          "condition": { "type": ["string", "null"] },
                          "price": { "type": "number", "description": "Listed price in USD" },
                          "msrp": { "type": ["number", "null"] },
                          "mileage": { "type": "integer" },
                          "exteriorColor": { "type": ["string", "null"] },
                          "interiorColor": { "type": ["string", "null"] },
                          "transmission": { "type": ["string", "null"] },
                          "engine": { "type": ["string", "null"] },
                          "fuelType": { "type": ["string", "null"] },
                          "drivetrain": { "type": ["string", "null"] },
                          "mpgCity": { "type": ["integer", "null"] },
                          "mpgHighway": { "type": ["integer", "null"] },
                          "doors": { "type": ["integer", "null"] },
                          "certifiedPreOwned": { "type": "boolean" },
                          "features": { "type": "array", "items": { "type": "string" } },
                          "photos": { "type": "array", "items": { "type": "string", "format": "uri" } },
                          "dealer": {
                            "type": "object",
                            "properties": {
                              "name": { "type": ["string", "null"] },
                              "city": { "type": ["string", "null"] },
                              "state": { "type": ["string", "null"] }
                            }
                          },
                          "distanceMiles": { "type": ["number", "null"], "description": "Distance from search coordinates in miles (only present when lat/lng provided)" },
                          "url": { "type": "string", "format": "uri", "description": "Direct link to vehicle listing on IQ Auto Deals" },
                          "listedAt": { "type": "string", "format": "date-time" }
                        }
                      }
                    },
                    "total": { "type": "integer", "description": "Total matching vehicles" },
                    "limit": { "type": "integer" },
                    "offset": { "type": "integer" },
                    "source": { "type": ["string", "null"] },
                    "description": { "type": ["string", "null"] }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request (invalid parameters)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["error"],
                  "properties": {
                    "error": { "type": "string" },
                    "details": { "type": ["string", "null"] }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["error"],
                  "properties": { "error": { "type": "string" } }
                }
              }
            }
          }
        }
      }
    }
  }
}